agregado funcionalidad Locacion

+++ PruebaUIs/Model/Parametros/Locacion.cs
@@ -21,8 +21,8 @@ namespace PruebaUIs.Model
         public int COD_CIU { get; set; }
         public int COD_BARR { get; set; }
         public string DES_DIR_LOC { get; set; }
-        public int VAL_ZLOC_ALT { get; set; }
-        public int VAL_ZLOC_SUP { get; set; }
+        public decimal VAL_ZLOC_ALT { get; set; }
+        public decimal VAL_ZLOC_SUP { get; set; }
         public int VAL_COB_ESP { get; set; }
         public string COD_USER { get; set; }
         public DateTime FEC_ABM { get; set; }
@@ -30,7 +30,7 @@ namespace PruebaUIs.Model
         public Locacion(bool fLG_DEP_CEN,
             string dES_LOC, DateTime fEC_CREA, string dES_NOM_ENC,
             bool fLG_LOC_VIR, string cOD_PAIS, int cOD_DPTO, int cOD_CIU,
-            int cOD_BARR, string dES_DIR_LOC, int vAL_ZLOC_ALT, int vAL_ZLOC_SUP,
+            int cOD_BARR, string dES_DIR_LOC, decimal vAL_ZLOC_ALT, decimal vAL_ZLOC_SUP,
             int vAL_COB_ESP, string cOD_USER, DateTime fEC_ABM)
         {
             FLG_DEP_CEN = fLG_DEP_CEN;
@@ -52,7 +52,7 @@ namespace PruebaUIs.Model
         public Locacion(bool fLG_DEP_CEN, string cOD_LOC,
             string dES_LOC, DateTime fEC_CREA, string dES_NOM_ENC,
             bool fLG_LOC_VIR, string cOD_PAIS, int cOD_DPTO, int cOD_CIU,
-            int cOD_BARR, string dES_DIR_LOC, int vAL_ZLOC_ALT, int vAL_ZLOC_SUP,
+            int cOD_BARR, string dES_DIR_LOC, decimal vAL_ZLOC_ALT, decimal vAL_ZLOC_SUP,
             int vAL_COB_ESP, string cOD_USER, DateTime fEC_ABM)
         {
             FLG_DEP_CEN = fLG_DEP_CEN;
